Add comment trigger to refresh translations in PR mode (#18)
