+++ README.md
@@ -26,6 +26,42 @@ GitHub Action to translate localization files using your [Weglot](https://weglot
     languages: fr,de,es
 ```
 
+### Refresh translations on demand (comment trigger)
+
+If you update translations directly in the Weglot dashboard and want to pull those changes into the open PR without pushing new source-file commits, you can comment `/update` on the PR. The action re-fetches translations from Weglot and pushes a new commit if anything changed, then replies with the result.
+
+Add `issue_comment` as a second trigger — the action handles the `/update` check itself:
+
+```yaml
+name: Translate
+
+on:
+  push:
+    branches: [main]
+    paths:
+      - locales/en.json
+  issue_comment:
+    types: [created]
+
+jobs:
+  translate:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      issues: write
+      pull-requests: write
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0   # full history needed for git rebase in pr mode
+      - uses: weglot/translate-action@v1
+        with:
+          api-key: ${{ secrets.WEGLOT_API_KEY }}
+          source-path: locales/en.json
+          output-mode: pr
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+```
+
 ### Create a PR with translations
 
 In **`output-mode: pr`**, the action uses a deterministic branch name `translations/<16-char-hex>` derived from your **`source-path`**, **`output-dir`**, and **API key** (hashed; the key is never placed in the branch name). **Target languages are not part of the hash**, so adding a language in Weglot or changing the `languages` input updates the same branch and the same open PR. Use different `source-path` and/or `output-dir` (or a different Weglot project key) for a second job in the same repository.
@@ -47,8 +83,6 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
 ```
 
-JavaScript actions do not receive `GITHUB_TOKEN` in the environment automatically—you must pass **`github-token`** (usually `${{ secrets.GITHUB_TOKEN }}`, same as `github.token`).
-
 ## Inputs
 
 | Input | Required | Default | Description |
@@ -58,7 +92,7 @@ JavaScript actions do not receive `GITHUB_TOKEN` in the environment automaticall
 | `output-dir` | No | (same as source) | Directory for translated files. |
 | `output-mode` | No | `files` | `files` = write to disk; `pr` = push to deterministic branch `translations/<hash>` and open or update one PR per stream (see above). |
 | `languages` | No | (from project) | Comma-separated target codes (e.g. `fr,de`). If empty, uses all languages from your Weglot project. |
-| `github-token` | When `output-mode: pr` | - | Token for the GitHub API and `git push`. Use `${{ secrets.GITHUB_TOKEN }}`. Job needs `contents: write` and `pull-requests: write`. |
+| `github-token` | When `output-mode: pr` | - | Token for the GitHub API and `git push`. Use `${{ secrets.GITHUB_TOKEN }}`. Job needs `contents: write`, `issues: write`, and `pull-requests: write`. |
 
 ## Outputs
