+++ eslint.config.mjs
@@ -16,6 +16,7 @@ export default tseslint.config(
       },
     },
     rules: {
+      "curly": ["error", "all"],
       "@typescript-eslint/no-unused-vars": [
         "error",
         { argsIgnorePattern: "^_" },
