+++ src/constants.ts
@@ -1,3 +1,4 @@
 export const CDN_BASE = "https://cdn.weglot.com";
 export const API_BASE = "https://cdn-api-weglot.com";
 export const API_MAX_LENGTH = 600;
+export const UPDATE_COMMENT_TRIGGER = "/update";
